# docs/swagger.yaml
definitions:
  common.RecordEventType:
    enum:
    - consent
    - pairing
    - issued
    - presented
    - revoked
    - presentationRequest
    - device.connection
    type: string
    x-enum-varnames:
    - Consent
    - Pairing
    - Issued
    - Presented
    - Revoked
    - PresentationRequest
    - DeviceConnection
  common.ServerErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.CreateDidPayload:
    properties:
      keyType:
        $ref: '#/definitions/types.KeyType'
    type: object
...
  handlers.DID:
    properties:
      detail:
        type: string
      did:
        type: string
      id:
        type: string
      timestamp:
        type: string
    type: object
  handlers.Event:
    properties:
      event:
        type: string
      timestamp:
        type: string
      type:
        $ref: '#/definitions/common.RecordEventType'
      userId:
        type: string
    type: object
...
  handlers.ListDIDResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/handlers.DID'
        type: array
    type: object
  handlers.ListHistoryOutput:
    properties:
      events:
        items:
          $ref: '#/definitions/handlers.Event'
        type: array
    type: object
...
  services.DIDCommConnection:
    properties:
      added:
        type: string
      eventType:
        type: string
      group:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
      protocol:
        type: string
      recipientDids:
        items:
          type: string
        type: array
      remoteDid:
        type: string
      routingKey:
        type: string
      topic:
        type: string
    type: object
...
  types.KeyType:
    enum:
    - aes256-gcm96
    - ed25519
    - ecdsa-p256
    - ecdsa-p384
    - ecdsa-p521
    - rsa-2048
    - rsa-3072
    - rsa-4096
    - keyValue
    type: string
    x-enum-varnames:
    - Aes256GCM
    - Ed25519
    - Ecdsap256
    - Ecdsap384
    - Ecdsap521
    - Rsa2048
    - Rsa3072
    - Rsa4096
    - KeyValue
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the personal credential manager cloud services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Account service API
  version: "1.0"
paths:
...
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a device
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Delete a device
      tags:
      - device
  /devices/block/{id}:
    post:
      consumes:
      - application/json
      description: Blocks a device
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
      summary: Block a device
      tags:
      - device
  /devices/link:
    get:
      consumes:
      - application/json
      description: Links a device
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Link
          schema:
            type: string
      summary: Link a device
      tags:
      - device
  /devices/list:
    get:
      consumes:
      - application/json
      description: Lists all devices
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.DIDCommConnection'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: List devices
      tags:
      - device
  /history/list:
    get:
      consumes:
      - application/json
      description: Lists all history events
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListHistoryOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: List history
      tags:
      - history
  /kms/did/{kid}:
    delete:
      consumes:
      - application/json
      description: Deletes a DID
      parameters:
      - description: DID ID
        in: path
        name: kid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Delete a DID
      tags:
      - kms
  /kms/did/create:
    post:
      consumes:
      - application/json
      description: Creates a DID
      parameters:
      - description: DID payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDidPayload'
      produces:
      - application/json
      responses:
        "200":
          description: DID ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Create a DID
      tags:
      - kms
  /kms/did/list:
    get:
      consumes:
      - application/json
      description: Lists all DIDs
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListDIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: List all DIDs
      tags:
      - kms
  /kms/keyTypes:
    get:
      consumes:
      - application/json
      description: Retrieves the supported keys algorithms
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.KeyType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get supported keys algorithms
      tags:
      - kms
...
swagger: "2.0"